Carrito terminado V4

+++ app/src/main/java/com/upao/elmochicaapp/models/Product.kt
@@ -3,6 +3,6 @@ package com.upao.elmochicaapp.models
 data class Product(
     val productName: String,
     val description: String,
-    val price: Double,
+    val price: Int,
     val image: Int // Puedes agregar un campo para el ID del recurso de imagen
 )
